SQL Server API/requirements.txt
fastapi
uvicorn
pymssql
python-dotenv